--- src/main/java/core/english/mse2023/handler/impl/AssignRoleHandler.java
@@ -1,43 +1,0 @@
-package core.english.mse2023.handler.impl;
-
-import core.english.mse2023.aop.annotation.handler.AdminRole;
-import core.english.mse2023.aop.annotation.handler.AllRoles;
-import core.english.mse2023.aop.annotation.handler.TextCommandType;
-import core.english.mse2023.component.ReplyKeyboardMaker;
-import core.english.mse2023.constant.ButtonCommand;
-import core.english.mse2023.handler.Handler;
-import lombok.AllArgsConstructor;
-import org.springframework.stereotype.Component;
-import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
-import org.telegram.telegrambots.meta.api.objects.Update;
-import org.telegram.telegrambots.meta.api.objects.commands.BotCommand;
-
-import java.util.List;
-
-@Component
-@AllRoles
-@TextCommandType
-@AllArgsConstructor
-public class AssignRoleHandler implements Handler {
-
-    private static final String MESSAGE = "Вы перешли в раздел НАЗНАЧИТЬ РОЛЬ";
-
-    private final ReplyKeyboardMaker replyKeyboardMaker;
-
-    @Override
-    public List<BotApiMethod<?>> handle(Update update) {
-
-        return List.of(SendMessage.builder()
-                .chatId(update.getMessage().getChatId().toString())
-                .text(MESSAGE)
-                .replyMarkup(replyKeyboardMaker.getAssignRoleKeyboard())
-                .build());
-    }
-
-    @Override
-    public BotCommand getCommandObject() {
-        return ButtonCommand.ASSIGN_ROLE;
-    }
-}
-
