--- src/main/java/core/english/mse2023/handler/impl/StartHandler.java
@@ -1,48 +1,0 @@
-package core.english.mse2023.handler.impl;
-
-import core.english.mse2023.aop.annotation.handler.AllRoles;
-import core.english.mse2023.aop.annotation.handler.TextCommandType;
-import core.english.mse2023.component.ReplyKeyboardMaker;
-import core.english.mse2023.constant.ButtonCommand;
-import core.english.mse2023.handler.Handler;
-import core.english.mse2023.model.User;
-import core.english.mse2023.service.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
-import org.telegram.telegrambots.meta.api.objects.Update;
-import org.telegram.telegrambots.meta.api.objects.commands.BotCommand;
-
-import java.util.List;
-
-@Component
-@TextCommandType
-@AllRoles
-@RequiredArgsConstructor
-public class StartHandler implements Handler {
-
-    private static final String GREETING = "Добро пожаловать, %s. Ваша роль: %s";
-
-
-    private final UserService service;
-
-    private final ReplyKeyboardMaker replyKeyboardMaker;
-
-
-    @Override
-    public List<BotApiMethod<?>> handle(Update update) {
-        User user = service.getUserOrCreateNewOne(update);
-
-        return List.of(SendMessage.builder()
-                .chatId(update.getMessage().getChatId().toString())
-                .text(String.format(GREETING, user.getName(), user.getRole()))
-                .replyMarkup(replyKeyboardMaker.getMainMenuKeyboard())
-                .build());
-    }
-
-    @Override
-    public BotCommand getCommandObject() {
-        return ButtonCommand.START;
-    }
-}
