+++ src/main/java/core/english/mse2023/handler/impl/menu/inline/GetMoreLessonInfoHandler.java
@@ -1,18 +1,21 @@
-package core.english.mse2023.handler.impl;
+package core.english.mse2023.handler.impl.done;
 
 import core.english.mse2023.aop.annotation.handler.AllRoles;
 import core.english.mse2023.aop.annotation.handler.InlineButtonType;
 import core.english.mse2023.component.InlineKeyboardMaker;
+import core.english.mse2023.component.MessageTextMaker;
 import core.english.mse2023.constant.InlineButtonCommand;
 import core.english.mse2023.dto.InlineButtonDTO;
 import core.english.mse2023.encoder.InlineButtonDTOEncoder;
 import core.english.mse2023.handler.Handler;
 import core.english.mse2023.model.Lesson;
+import core.english.mse2023.model.dictionary.UserRole;
 import core.english.mse2023.service.LessonService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
-import org.telegram.telegrambots.meta.api.methods.updatingmessages.EditMessageReplyMarkup;
+import org.telegram.telegrambots.meta.api.methods.ParseMode;
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Update;
 import org.telegram.telegrambots.meta.api.objects.commands.BotCommand;
 
@@ -23,30 +26,36 @@ import java.util.UUID;
 @InlineButtonType
 @AllRoles
 @RequiredArgsConstructor
-public class GetMoreSubscriptionInfoHandler implements Handler {
+public class GetMoreLessonInfoHandler implements Handler {
+
+    private final MessageTextMaker messageTextMaker;
 
     private final LessonService lessonService;
 
     private final InlineKeyboardMaker inlineKeyboardMaker;
 
     @Override
-    public List<BotApiMethod<?>> handle(Update update) {
+    public List<BotApiMethod<?>> handle(Update update, UserRole userRole) {
 
         InlineButtonDTO buttonData = InlineButtonDTOEncoder.decode(update.getCallbackQuery().getData());
 
-        List<Lesson> lessons = lessonService.getAllLessonsForSubscription(UUID.fromString(buttonData.getData()));
+        Lesson lesson = lessonService.getLessonById(UUID.fromString(buttonData.getData()));
 
-        EditMessageReplyMarkup editMessageReplyMarkup = EditMessageReplyMarkup.builder()
+
+        SendMessage message = SendMessage.builder()
                 .chatId(update.getCallbackQuery().getMessage().getChatId().toString())
-                .messageId(update.getCallbackQuery().getMessage().getMessageId())
-                .replyMarkup(inlineKeyboardMaker.getSubscriptionLessonsMenu(lessons, buttonData.getData()))
+                .text(messageTextMaker.lessonInfoPatternMessageText(lesson))
                 .build();
 
-        return List.of(editMessageReplyMarkup);
+        message.setParseMode(ParseMode.MARKDOWNV2);
+
+        message.setReplyMarkup(inlineKeyboardMaker.getLessonMainMenuInlineKeyboard(lesson.getId().toString(), lesson.getStatus(), userRole));
+
+        return List.of(message);
     }
 
     @Override
     public BotCommand getCommandObject() {
-        return InlineButtonCommand.GET_MORE_SUBSCRIPTION_INFO;
+        return InlineButtonCommand.GET_MORE_LESSON_INFO;
     }
 }
