+++ src/main/java/core/english/mse2023/handler/impl/menu/inline/GetMoreSubscriptionInfoHandler.java
@@ -1,4 +1,4 @@
-package core.english.mse2023.handler.impl;
+package core.english.mse2023.handler.impl.done;
 
 import core.english.mse2023.aop.annotation.handler.AllRoles;
 import core.english.mse2023.aop.annotation.handler.InlineButtonType;
@@ -7,6 +7,9 @@ import core.english.mse2023.constant.InlineButtonCommand;
 import core.english.mse2023.dto.InlineButtonDTO;
 import core.english.mse2023.encoder.InlineButtonDTOEncoder;
 import core.english.mse2023.handler.Handler;
+import core.english.mse2023.model.Lesson;
+import core.english.mse2023.model.dictionary.UserRole;
+import core.english.mse2023.service.LessonService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
@@ -15,28 +18,36 @@ import org.telegram.telegrambots.meta.api.objects.Update;
 import org.telegram.telegrambots.meta.api.objects.commands.BotCommand;
 
 import java.util.List;
+import java.util.UUID;
 
 @Component
 @InlineButtonType
 @AllRoles
 @RequiredArgsConstructor
-public class MainMenuLessonHandler implements Handler {
+public class GetMoreSubscriptionInfoHandler implements Handler {
+
+    private final LessonService lessonService;
 
     private final InlineKeyboardMaker inlineKeyboardMaker;
 
     @Override
-    public List<BotApiMethod<?>> handle(Update update) {
+    public List<BotApiMethod<?>> handle(Update update, UserRole userRole) {
+
         InlineButtonDTO buttonData = InlineButtonDTOEncoder.decode(update.getCallbackQuery().getData());
-        return List.of(EditMessageReplyMarkup.builder()
-                        .chatId(update.getCallbackQuery().getMessage().getChatId().toString())
-                        .messageId(update.getCallbackQuery().getMessage().getMessageId())
-                        .replyMarkup(inlineKeyboardMaker.getLessonMainMenuInlineKeyboard(buttonData.getData()))
-                        .build()
-        );
+
+        List<Lesson> lessons = lessonService.getAllLessonsForSubscription(UUID.fromString(buttonData.getData()));
+
+        EditMessageReplyMarkup editMessageReplyMarkup = EditMessageReplyMarkup.builder()
+                .chatId(update.getCallbackQuery().getMessage().getChatId().toString())
+                .messageId(update.getCallbackQuery().getMessage().getMessageId())
+                .replyMarkup(inlineKeyboardMaker.getSubscriptionLessonsMenu(lessons, buttonData.getData()))
+                .build();
+
+        return List.of(editMessageReplyMarkup);
     }
 
     @Override
     public BotCommand getCommandObject() {
-        return InlineButtonCommand.MAIN_MENU_LESSON;
+        return InlineButtonCommand.GET_MORE_SUBSCRIPTION_INFO;
     }
 }
