+++ src/main/java/core/english/mse2023/handler/impl/menu/inline/MainMenuLessonHandler.java
@@ -1,4 +1,4 @@
-package core.english.mse2023.handler.impl;
+package core.english.mse2023.handler.impl.done;
 
 import core.english.mse2023.aop.annotation.handler.AllRoles;
 import core.english.mse2023.aop.annotation.handler.InlineButtonType;
@@ -7,6 +7,10 @@ import core.english.mse2023.constant.InlineButtonCommand;
 import core.english.mse2023.dto.InlineButtonDTO;
 import core.english.mse2023.encoder.InlineButtonDTOEncoder;
 import core.english.mse2023.handler.Handler;
+import core.english.mse2023.model.Lesson;
+import core.english.mse2023.model.dictionary.LessonStatus;
+import core.english.mse2023.model.dictionary.UserRole;
+import core.english.mse2023.service.LessonService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
@@ -15,29 +19,34 @@ import org.telegram.telegrambots.meta.api.objects.Update;
 import org.telegram.telegrambots.meta.api.objects.commands.BotCommand;
 
 import java.util.List;
+import java.util.UUID;
 
 @Component
 @InlineButtonType
 @AllRoles
 @RequiredArgsConstructor
-public class MainMenuSubscriptionHandler implements Handler {
+public class MainMenuLessonHandler implements Handler {
 
     private final InlineKeyboardMaker inlineKeyboardMaker;
 
+    private final LessonService lessonService;
 
     @Override
-    public List<BotApiMethod<?>> handle(Update update) {
+    public List<BotApiMethod<?>> handle(Update update, UserRole userRole) {
         InlineButtonDTO buttonData = InlineButtonDTOEncoder.decode(update.getCallbackQuery().getData());
 
+        Lesson lesson = lessonService.getLessonById(UUID.fromString(buttonData.getData()));
+
         return List.of(EditMessageReplyMarkup.builder()
-                .chatId(update.getCallbackQuery().getMessage().getChatId().toString())
-                .messageId(update.getCallbackQuery().getMessage().getMessageId())
-                .replyMarkup(inlineKeyboardMaker.getSubscriptionMainMenu(buttonData.getData()))
-                .build());
+                        .chatId(update.getCallbackQuery().getMessage().getChatId().toString())
+                        .messageId(update.getCallbackQuery().getMessage().getMessageId())
+                        .replyMarkup(inlineKeyboardMaker.getLessonMainMenuInlineKeyboard(buttonData.getData(), lesson.getStatus(), userRole))
+                        .build()
+        );
     }
 
     @Override
     public BotCommand getCommandObject() {
-        return InlineButtonCommand.MAIN_MENU_SUBSCRIPTION;
+        return InlineButtonCommand.MAIN_MENU_LESSON;
     }
 }
