+++ src/main/java/core/english/mse2023/handler/impl/menu/regular/StartHandler.java
@@ -1,11 +1,14 @@
-package core.english.mse2023.handler.impl;
+package core.english.mse2023.handler.impl.done;
 
 import core.english.mse2023.aop.annotation.handler.AllRoles;
 import core.english.mse2023.aop.annotation.handler.TextCommandType;
 import core.english.mse2023.component.ReplyKeyboardMaker;
 import core.english.mse2023.constant.ButtonCommand;
 import core.english.mse2023.handler.Handler;
-import lombok.AllArgsConstructor;
+import core.english.mse2023.model.User;
+import core.english.mse2023.model.dictionary.UserRole;
+import core.english.mse2023.service.UserService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
@@ -17,25 +20,30 @@ import java.util.List;
 @Component
 @TextCommandType
 @AllRoles
-@AllArgsConstructor
-public class FullStatisticsHandler implements Handler {
+@RequiredArgsConstructor
+public class StartHandler implements Handler {
 
-    private static final String MESSAGE = "Вы перешли в раздел ПОЛНАЯ СТАТИСТИКА ПО ШКОЛЕ";
+    private static final String GREETING = "Добро пожаловать, %s. Ваша роль: %s";
+
+
+    private final UserService service;
+
     private final ReplyKeyboardMaker replyKeyboardMaker;
 
+
     @Override
-    public List<BotApiMethod<?>> handle(Update update) {
+    public List<BotApiMethod<?>> handle(Update update, UserRole userRole) {
+        User user = service.getUserOrCreateNewOne(update);
 
         return List.of(SendMessage.builder()
                 .chatId(update.getMessage().getChatId().toString())
-                .text(MESSAGE)
-                .replyMarkup(replyKeyboardMaker.getFullStatisticsKeyboard())
+                .text(String.format(GREETING, user.getName(), user.getRole()))
+                .replyMarkup(replyKeyboardMaker.getMainMenuKeyboard(userRole))
                 .build());
     }
 
     @Override
     public BotCommand getCommandObject() {
-        return ButtonCommand.SCHOOL_STATISTICS;
+        return ButtonCommand.START;
     }
 }
-
