+++ src/main/java/core/english/mse2023/handler/impl/menu/regular/ToMainMenuHandler.java
@@ -1,10 +1,11 @@
-package core.english.mse2023.handler.impl;
+package core.english.mse2023.handler.impl.menu;
 
 import core.english.mse2023.aop.annotation.handler.AllRoles;
 import core.english.mse2023.aop.annotation.handler.TextCommandType;
 import core.english.mse2023.component.ReplyKeyboardMaker;
 import core.english.mse2023.constant.ButtonCommand;
 import core.english.mse2023.handler.Handler;
+import core.english.mse2023.model.dictionary.UserRole;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
@@ -18,24 +19,23 @@ import java.util.List;
 @TextCommandType
 @AllRoles
 @AllArgsConstructor
-public class StudentStatisticsMenuHandler implements Handler {
+public class ToMainMenuHandler implements Handler {
 
-    private static final String MESSAGE = "Вы перешли в раздел СТАТИСТИКА ПО УЧЕНИКУ";
+    private static final String MESSAGE = "Вы вернулись в ГЛАВНОЕ МЕНЮ";
     private final ReplyKeyboardMaker replyKeyboardMaker;
 
     @Override
-    public List<BotApiMethod<?>> handle(Update update) {
+    public List<BotApiMethod<?>> handle(Update update, UserRole userRole) {
 
         return List.of(SendMessage.builder()
                         .chatId(update.getMessage().getChatId().toString())
                         .text(MESSAGE)
-                        .replyMarkup(replyKeyboardMaker.getStudentStatisticsKeyboard())
+                        .replyMarkup(replyKeyboardMaker.getMainMenuKeyboard(userRole))
                         .build());
     }
 
     @Override
     public BotCommand getCommandObject() {
-        return ButtonCommand.STUDENT_STATISTICS;
+        return ButtonCommand.TO_MAIN_MENU;
     }
 }
-
